refactor: Last edit placement in flow breadcrumbs
